refactor(workspace): generalize attention focus handling
