+++ packages/workspace/src/front/attention/index.ts
@@ -12,4 +12,5 @@ export type {
   WorkspaceAttentionContextValue,
   WorkspaceAttentionProviderProps,
   WorkspaceAttentionSessionBadge,
+  WorkspaceAttentionFocusMetadata,
 } from "./WorkspaceAttentionProvider"
